CHROMIUM: sched: tune: allow nested schedtune groups
Adapted to 4.14.290 by berbascum

Let's remove restriction that schedtune groups can not be nested. Note
that they are still completely independent even if they are nested, but
at least we can build proper hierarchies now.

We also extend maximum number of schedtune groups to 16 for now.

CHROMIUM Note: Apparently the plan is to drop schedtune in favor of
utilization clamping. Future kernel rebases should hopefully be able
to drop this patch once utilization clamping is accepted upstream and
we've moved to it. Check out the second bug in the BUG= field.